bot/requirements.txt
discord.py
pytz
PyYAML
